Updated invitations

+++ src/Modules/Stakeholders/Explorer.Stakeholders.Core/UseCases/ClubMemberManagementService.cs
@@ -76,10 +76,10 @@ public class ClubMemberManagementService : IClubMemberManagementService
             var membership = _clubMembershipRepository.Get(membershipId);
             var club = _clubRepository.Get(membership.ClubId);
 
-            //if (club.OwnerId != userId)
-            //{
-            //    return Result.Fail(FailureCode.InvalidArgument).WithError(FailureCode.InvalidArgument);
-            //}
+            if (club.OwnerId != userId)
+            {
+                return Result.Fail(FailureCode.InvalidArgument).WithError(FailureCode.InvalidArgument);
+            }
 
             _clubMembershipRepository.Delete(membershipId);
